server/requirements/dev.txt
-r ./prod.txt

django-debug-toolbar==2.2
flake8==3.8.4
